🎨 [NFT Widget] Restructure code
